Backend/requirements.txt
Django==5.0.8
psycopg[binary]==3.2.1
python-dotenv==1.0.1
gunicorn==22.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers
gtts==2.3.2